+++ src/html.rs
@@ -10,6 +10,17 @@ where
     for event in iterator {
         match event {
             Begin(tag) => match tag {
+                FencedCodeBlock(info) => {
+                    let language = info.split(' ').next().unwrap();
+                    if language.is_empty() {
+                        writer.push_str("<pre><code>");
+                    } else {
+                        writer.push_str(&format!(
+                            r#"<pre><code class="language-{language}">"#,
+                            language = language
+                        ));
+                    }
+                }
                 Heading(level) => {
                     writer.push_str(&format!("<h{}>", level as usize));
                 }
@@ -21,11 +32,11 @@ where
                 }
             },
             End(tag) => match tag {
+                FencedCodeBlock(_) | IndentedCodeBlock => {
+                    writer.push_str("</code></pre>\n");
+                }
                 Heading(level) => {
                     writer.push_str(&format!("</h{}>\n", level as usize));
-                }
-                IndentedCodeBlock => {
-                    writer.push_str("</code></pre>\n");
                 }
                 Paragraph => {
                     writer.push_str("</p>\n");