+++ src/types/block.rs
@@ -1,14 +1,15 @@
 use crate::types::HeadingLevel;
 
 #[derive(Clone, Copy, Debug)]
-pub struct Block {
+pub struct Block<'a> {
     pub begin: usize,
     pub end: usize,
-    pub kind: BlockKind,
+    pub kind: BlockKind<'a>,
 }
 
 #[derive(Clone, Copy, Debug)]
-pub enum BlockKind {
+pub enum BlockKind<'a> {
+    FencedCodeBlock(&'a str),
     Heading(HeadingLevel),
     IndentedCodeBlock,
     Paragraph,