+++ src/types/event.rs
@@ -2,8 +2,8 @@ use crate::types::Tag;
 
 #[derive(Debug, PartialEq)]
 pub enum Event<'a> {
-    Begin(Tag),
-    End(Tag),
+    Begin(Tag<'a>),
+    End(Tag<'a>),
     Text(&'a str),
     ThematicBreak,
 }